backend/requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic
psycopg2-binary
